# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateBrandRequest:
    properties:
      equipment_type_id:
        type: integer
      name:
        type: string
    required:
    - equipment_type_id
    - name
    type: object
...
  dto.CreateEquipmentModelRequest:
    properties:
      brand_id:
        type: integer
      equipment_type_id:
        type: integer
      name:
        type: string
    required:
    - brand_id
    - equipment_type_id
    - name
    type: object
...
  dto.CreateEquipmentTypeRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  dto.UpdateBrandRequest:
    properties:
      equipment_type_id:
        type: integer
      name:
        type: string
    required:
    - equipment_type_id
    - name
    type: object
...
  dto.UpdateEquipmentModelRequest:
    properties:
      brand_id:
        type: integer
      equipment_type_id:
        type: integer
      name:
        type: string
    required:
    - brand_id
    - equipment_type_id
    - name
    type: object
...
  dto.UpdateEquipmentTypeRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      success:
        type: boolean
    type: object
  response.Pagination:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@speckit.com
    name: API Support
  description: API for managing technical services, clients, equipment, work orders,
    and materials.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://swagger.io/terms/
  title: Service Management API
  version: 1.0.0
paths:
...
  /api/v1/catalog/brands:
    post:
      consumes:
      - application/json
      description: Create a new brand in the catalog
      parameters:
      - description: Brand data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBrandRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Create brand
      tags:
      - Catalog
  /api/v1/catalog/brands/{id}:
    delete:
      description: Delete a brand by ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Delete brand
      tags:
      - Catalog
    get:
      description: Retrieve a single brand by ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get brand by ID
      tags:
      - Catalog
    put:
      consumes:
      - application/json
      description: Update an existing brand
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated brand data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBrandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Update brand
      tags:
      - Catalog
  /api/v1/catalog/brands/by-type/{typeId}:
    get:
      description: Retrieve all brands for a given equipment type
      parameters:
      - description: Equipment type ID
        in: path
        name: typeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: List brands by type
      tags:
      - Catalog
  /api/v1/catalog/models:
    post:
      consumes:
      - application/json
      description: Create a new equipment model in the catalog
      parameters:
      - description: Equipment model data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEquipmentModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Create equipment model
      tags:
      - Catalog
  /api/v1/catalog/models/{id}:
    delete:
      description: Delete an equipment model by ID
      parameters:
      - description: Equipment model ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Delete equipment model
      tags:
      - Catalog
    get:
      description: Retrieve a single equipment model by ID
      parameters:
      - description: Equipment model ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get equipment model
      tags:
      - Catalog
    put:
      consumes:
      - application/json
      description: Update an existing equipment model
      parameters:
      - description: Equipment model ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated equipment model data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEquipmentModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Update equipment model
      tags:
      - Catalog
  /api/v1/catalog/models/by-brand/{brandId}:
    get:
      description: Retrieve all equipment models for a given brand
      parameters:
      - description: Brand ID
        in: path
        name: brandId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: List models by brand
      tags:
      - Catalog
  /api/v1/catalog/types:
    get:
      description: Retrieve all equipment types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: List equipment types
      tags:
      - Catalog
    post:
      consumes:
      - application/json
      description: Create a new equipment type in the catalog
      parameters:
      - description: Equipment type data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEquipmentTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Create equipment type
      tags:
      - Catalog
  /api/v1/catalog/types/{id}:
    delete:
      description: Delete an equipment type by ID
      parameters:
      - description: Equipment type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Delete equipment type
      tags:
      - Catalog
    get:
      description: Retrieve a single equipment type by ID
      parameters:
      - description: Equipment type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get equipment type
      tags:
      - Catalog
    put:
      consumes:
      - application/json
      description: Update an existing equipment type
      parameters:
      - description: Equipment type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated equipment type data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEquipmentTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Update equipment type
      tags:
      - Catalog
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"